refactor: Reposicionando o enum para a pasta model
