+++ back-end/src/main/java/dev/gui/processo_sergipetec/model/CarroModel.java
@@ -1,6 +1,5 @@
 package dev.gui.processo_sergipetec.model;
 
-import dev.gui.processo_sergipetec.TipoCombustível;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
