+++ back-end/src/main/java/dev/gui/processo_sergipetec/model/TipoCombustível.java
@@ -1,4 +1,4 @@
-package dev.gui.processo_sergipetec;
+package dev.gui.processo_sergipetec.model;
 
 public enum TipoCombustível {
     GASOLINA, ETANOL, DIESEL, FLEX
